+++ example/__init__.py
@@ -1,5 +1,0 @@
-import logging
-import typing as tp
-
-logger = logging.getLogger(__name__)
-