+++ snakehouse/__init__.py
@@ -2,4 +2,4 @@ from .build import build
 from .multibuild import Multibuild
 from .faster_builds import monkey_patch_parallel_compilation
 
-__version__ = '1.2.4a2'
+__version__ = '1.3a1'